requirements.txt
amqp
asgiref
async-timeout
billiard
cachetools
celery
certifi
cffi
chardet
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
coverage
cron-descriptor
cryptography
defer
defusedxml
distro-info
Django
django-better-admin-arrayfield
django-celery-beat
django-ckeditor
django-cors-headers
django-debug-toolbar
django-environ
django-js-asset
django-redis
django-timezone-field
djangorestframework
djangorestframework-simplejwt
drf-extra-fields
drf-yasg
filetype
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
googleapis-common-protos
gunicorn
httplib2
idna
inflection
jeepney
kombu
louis
oauthlib
packaging
pillow
pip-autoremove
prompt_toolkit
proto-plus
protobuf
psycopg2-binary
pyasn1
pyasn1_modules
pycparser
PyJWT
pyparsing
python-crontab
python-dateutil
python-http-client
python3-openid
pytz
PyYAML
redis
reportlab
requests
requests-oauthlib
resolver
rsa
SecretStorage
selector
sendgrid
setuptools
six
social-auth-app-django
social-auth-core
sqlparse
starkbank-ecdsa
systemd-python
texttable
tk
typing_extensions
tzdata
uritemplate
urllib3
vine
wcwidth
